--- src/characters/character.repository.ts
@@ -1,7 +1,0 @@
-import { Character } from './Character'
-
-export const CHARACTER_REPOSITORY_PROVIDER = 'CHARACTER_REPOSITORY'
-
-export interface CharacterRepository {
-  getById(id: string): Promise<Character | null>
-}
