+++ src/characters/domain/character.errors.ts
@@ -1,0 +1,3 @@
+export enum CharacterErrorCode {
+  NOT_FOUND = 'CH-000-001',
+}
