--- src/characters/error/character.errors.ts
@@ -1,3 +1,0 @@
-export enum CharacterErrorCode {
-  NOT_FOUND = 'CH-000-001',
-}
