+++ src/config/app.module.ts
@@ -1,15 +1,14 @@
 import { Module } from '@nestjs/common'
-import { PingController } from '../controllers/ping/ping.controller'
 import { ConfigModule } from '@nestjs/config'
 import configuration from './env.config'
-import { CHARACTER_REPOSITORY_PROVIDER } from 'src/characters/character.repository'
-import { PiratesController } from 'src/controllers/pirate.controller'
-import { SCRAPER_REPOSITORY_PROVIDER } from 'src/scraper/scraper'
-import AsideJSDomScraper from 'src/scraper/aside.jsdom.scraper'
-import CharacterScraperRepository from 'src/characters/character.scraper.repository'
-import CharacterService from 'src/characters/character.service'
+import { CHARACTER_REPOSITORY_PROVIDER } from 'src/characters/domain/character.repository'
+import { CharacterController } from 'src/characters/infra/rest/character.controller'
+import { SCRAPER_REPOSITORY_PROVIDER } from 'src/scraper/domain/scraper'
+import AsideJSDomScraper from 'src/scraper/infra/jsdom/aside.jsdom.scraper'
+import CharacterScraperRepository from 'src/characters/infra/db/scraper/character.scraper.repository'
+import CharacterService from 'src/characters/application/character.service'
 import { APP_FILTER } from '@nestjs/core'
-import { HttpErrorFilter } from 'src/middlewares/http-error.filter'
+import { HttpErrorFilter } from 'src/shared/error/infra/rest/http-error.filter'
 
 @Module({
   imports: [
@@ -34,6 +33,6 @@ import { HttpErrorFilter } from 'src/middlewares/http-error.filter'
     },
     CharacterService,
   ],
-  controllers: [PingController, PiratesController],
+  controllers: [CharacterController],
 })
 export class AppModule {}
