--- src/controllers/ping/ping.controller.ts
@@ -1,9 +1,0 @@
-import { Controller, Get } from '@nestjs/common'
-
-@Controller({ path: '/ping' })
-export class PingController {
-  @Get()
-  async getHello() {
-    return { message: 'Pong' }
-  }
-}
